+++ libs/trade/executor/base_executor.py
@@ -14,9 +14,9 @@ class BaseExecutor(ABC):
 
         Args:
             symbol (str): The symbol of the market, ie. 'ETH/USD'.
-            quote_amount (float): The amount to buy in quote currency.
-            asset_price (float): The target buy price, quote per base.
-            slippage (float): The percentage off asset_price the market
+            quote_amount (Decimal): The amount to buy in quote currency.
+            asset_price (Decimal): The target buy price, quote per base.
+            slippage (Decimal): The percentage off asset_price the market
                 buy will tolerate.
         """
         pass
@@ -28,9 +28,9 @@ class BaseExecutor(ABC):
 
         Args:
             symbol (str): The symbol of the market, ie. 'ETH/USD'.
-            asset_price (float): The price, quote per base.
-            asset_amount (float): The amount of the asset to be sold.
-            slippage (float): The percentage off asset_price the market
+            asset_price (Decimal): The price, quote per base.
+            asset_amount (Decimal): The amount of the asset to be sold.
+            slippage (Decimal): The percentage off asset_price the market
                 buy will tolerate.
         """
         pass
@@ -41,8 +41,8 @@ class BaseExecutor(ABC):
 
         Args:
             symbol (str): The symbol of the market, ie. 'ETH/USD'.
-            asset_amount (float): The amount of asset to be bought.
-            asset_price (float); The target buy price, quote per base.
+            asset_amount (Decimal): The amount of asset to be bought.
+            asset_price (Decimal); The target buy price, quote per base.
         """
         pass
 
@@ -52,7 +52,7 @@ class BaseExecutor(ABC):
 
         Args:
             symbol (str): The symbol of the market, ie. 'ETH/USD'.
-            asset_amount (float): The amount of asset to be sold.
-            asset_price (float); The target sell price, quote per base.
+            asset_amount (Decimal): The amount of asset to be sold.
+            asset_price (Decimal); The target sell price, quote per base.
         """
         pass
